multiple admins

+++ classes_and_fee/classes_and_fee_delete.php
@@ -6,7 +6,7 @@
 		$id = $_POST['id'];
 		$conn = $pdo->open();
 		try{
-			$stmt = $conn->prepare("DELETE FROM classes_and_fee WHERE classes_and_fee_id=:id");
+			$stmt = $conn->prepare("DELETE FROM classes_and_fee WHERE classes_and_fee_id=:id AND classes_and_fee_school_id=" . $_SESSION['admin_school_id'] . " ");
 			$stmt->execute(['id'=>$id]);
 
 			$_SESSION['success'] = 'Class and fee deleted successfully';
